+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # MiniReel Player
 
+* *[Beta20.rc1]*
+  * Collateral assets now live in their own GH repo
+  * [FEATURE]: Added support for text cards
+
 ## Beta19 (July 22, 2014)
+* *[Beta19.rc6]*
+  * Made txt on fallback display ad easier to read
+* *[/Beta19.rc6]*
+
 * *[Beta19.rc5]*
   * [FIX]: Reset 'preloaded' property on static ad cards when MR is reset
 * *[/Beta19.rc5]*